クリップボードコピー、SyncDelay整数化、FB整数化、OPM FREQ整数化、OPZX3ピッチエンベロープをオペレータ単位に変更

+++ 2686V/Source/Synth/Opna/Params.h
@@ -57,7 +57,7 @@ struct OpnaParams
     float lfoAmSmRt = 0.005f;
 
     // LFO Sync Delay
-    float lfoSyncDelay = 0.0f;
+    int lfoSyncDelay = 0;
 
     // --- OPLL Preset ---
     int opllPreset = 0; // OPLL Preset Instrument Index
